Refactor App to use layouts
