+++ src/components/layouts/BlankLayout.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+  <div>
+    
+  </div>
+</template>
